--- src/main/java/com/balamut/yejournal/Admins.java
@@ -1,21 +1,0 @@
-package com.balamut.yejournal;
-
-
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-
-import java.util.UUID;
-
-@Entity
-public class Admins {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.SEQUENCE)
-    private Long id;
-
-    private UUID school;
-
-    private UUID admin;
-}
