--- src/main/java/com/balamut/yejournal/StudentRepository.java
@@ -1,12 +1,0 @@
-package com.balamut.yejournal;
-
-import org.springframework.data.repository.CrudRepository;
-
-import java.util.UUID;
-
-public interface StudentRepository extends CrudRepository<Student, UUID> {
-
-    boolean existsByUuid(UUID uuid);
-
-    Student getByUuid(UUID uuid);
-}
