+++ src/main/java/com/balamut/yejournal/entities/Lesson.java
@@ -1,0 +1,29 @@
+package com.balamut.yejournal.entities;
+
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import java.util.UUID;
+
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+@Data
+@Entity
+public class Lesson {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.UUID)
+    private UUID id;
+    private UUID classroomId;
+    private UUID teacherId;
+    private String name;
+    private Long begin;
+    private Long end;
+}
